@@ -1,10 +1,10 @@
 module github.com/r0bj/gke-events-notifier
 
-go 1.16
+go 1.23.4
+
+require github.com/alecthomas/kingpin/v2 v2.4.0
 
 require (
-	github.com/alecthomas/template v0.0.0-20190718012654-fb15b899a751 // indirect
-	github.com/alecthomas/units v0.0.0-20210208195552-ff826a37aa15 // indirect
-	github.com/sirupsen/logrus v1.8.1 // indirect
-	gopkg.in/alecthomas/kingpin.v2 v2.2.6 // indirect
+	github.com/alecthomas/units v0.0.0-20211218093645-b94a6e3cc137 // indirect
+	github.com/xhit/go-str2duration/v2 v2.1.0 // indirect
 )
